# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nina Ofertas API",
  "version": "0.5.0"
 },
 "paths": {
...
  "/api/phones": {
   "get": {
    "tags": [
     "phones"
    ],
    "summary": "List Phones",
    "operationId": "list_phones_api_phones_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "-created_at",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaginatedResponse_PhoneResponse_"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "phones"
    ],
    "summary": "Create Phone",
    "operationId": "create_phone_api_phones_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/phones/{phone_id}": {
   "patch": {
    "tags": [
     "phones"
    ],
    "summary": "Update Phone",
    "operationId": "update_phone_api_phones__phone_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "phone_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Phone Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "phones"
    ],
    "summary": "Delete Phone",
    "operationId": "delete_phone_api_phones__phone_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "phone_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Phone Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/phones/{phone_id}/qrcode": {
   "get": {
    "tags": [
     "phones"
    ],
    "summary": "Get Qrcode",
    "operationId": "get_qrcode_api_phones__phone_id__qrcode_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "phone_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Phone Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QRCodeResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
  "/api/phones/{phone_id}/sync-groups": {
   "post": {
    "tags": [
     "phones"
    ],
    "summary": "Sync Groups",
    "operationId": "sync_groups_api_phones__phone_id__sync_groups_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "phone_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Phone Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CommandResponse"
        }
       }
      }
     },
     "401": {
      "description": "Não autenticado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "403": {
      "description": "Sem permissão",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "404": {
      "description": "Não encontrado",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "409": {
      "description": "Conflito (ex.: entidade em uso)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "422": {
      "description": "Dados inválidos",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "429": {
      "description": "Limite de requisições",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     },
     "500": {
      "description": "Erro interno",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorEnvelope"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CommandResponse": {
    "properties": {
     "command_id": {
      "type": "integer",
      "title": "Command Id"
     }
    },
    "type": "object",
    "required": [
     "command_id"
    ],
    "title": "CommandResponse"
   },
...
   "ErrorDetail": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "fields": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fields"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message"
    ],
    "title": "ErrorDetail"
   },
   "ErrorEnvelope": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorDetail"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorEnvelope"
   },
...
   "PaginatedResponse_PhoneResponse_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PhoneResponse"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "PaginatedResponse[PhoneResponse]"
   },
...
   "PhoneCreate": {
    "properties": {
     "label": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Label"
     },
     "number": {
      "type": "string",
      "title": "Number"
     },
     "evolution_instance": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Evolution Instance"
     }
    },
    "type": "object",
    "required": [
     "label",
     "number"
    ],
    "title": "PhoneCreate"
   },
   "PhoneResponse": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "number": {
      "type": "string",
      "title": "Number"
     },
     "evolution_instance": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evolution Instance"
     },
     "status": {
      "type": "string",
      "enum": [
       "connected",
       "disconnected",
       "banned",
       "unknown"
      ],
      "title": "Status"
     },
     "last_seen_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "label",
     "number",
     "evolution_instance",
     "status",
     "last_seen_at",
     "created_at"
    ],
    "title": "PhoneResponse"
   },
...
   "PhoneUpdate": {
    "properties": {
     "label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     },
     "evolution_instance": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Evolution Instance"
     }
    },
    "type": "object",
    "title": "PhoneUpdate"
   },
...
   "QRCodeResponse": {
    "properties": {
     "qrcode_base64": {
      "type": "string",
      "title": "Qrcode Base64"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "qrcode_base64",
     "expires_at"
    ],
    "title": "QRCodeResponse"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}